Refactor login, signup to form directory integration
